# doc/OpenAPI_swagger.yml
openapi: 3.0.0
info:
  title: Participium API
  description: API for the Participium application
  version: 0.1.0
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /reports:
    # create report endpoint
    post:
      tags: [Reports]
      summary: Create a new report
      description: Submit a new report (authenticated users only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Created - Report successfully submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Bad Request - Invalid input or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - User is not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Authentication Error"
                message: "You must be logged in to create a report"
    get:
      tags: [Reports]
      summary: List all reports
      description: >
        Returns the list of all reports. 
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - ASSIGNED
          description: Filter reports by status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - You don't have admin or municipality permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:    
...
    ReportCategory:
      type: string
      enum:
        - Water Supply – Drinking Water
        - Architectural Barriers
        - Sewer System
        - Public Lighting
        - Waste
        - Road Signs and Traffic Lights
        - Roads and Urban Furnishings
        - Public Green Areas and Playgrounds
        - Other
        - WATER_SUPPLY_DRINKING_WATER
        - ARCHITECTURAL_BARRIERS
        - SEWER_SYSTEM
        - PUBLIC_LIGHTING
        - WASTE
        - ROAD_SIGNS_TRAFFIC_LIGHTS
        - ROADS_URBAN_FURNISHINGS
        - PUBLIC_GREEN_AREAS_PLAYGROUNDS
        - OTHER
      example: "Public Lighting"
    ReportRequest:
      type: object
      required:
        - title
        - description
        - category
        - photos
        - latitude
        - longitude
      properties:
        title:
          type: string
          example: "Broken street light"
        description:
          type: string
          example: "The street light on Via Roma has been broken for a week"
        anonymous:
          type: boolean
          example: false
        category:
          $ref: '#/components/schemas/ReportCategory'
        photos:
          type: array
          items:
            type: string
            format: binary
          minItems: 1
          maxItems: 3
          description: Array of image files to upload (multipart/form-data)
        latitude:
          type: number
          format: float
          example: 45.4642
        longitude:
          type: number
          format: float 
          example: 9.1900
    Report:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Broken street light"
        description:
          type: string
          example: "The street light on Via Roma has been broken for a week"
        anonymous:
          type: boolean
          example: false
        category:
          $ref: '#/components/schemas/ReportCategory'
        photos:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 3
          example: 
            - "https://example.com/photos/report1_photo1.jpg"
            - "https://example.com/photos/report1_photo2.jpg"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-04T14:30:00Z"
        latitude:
          type: number
          format: float
          example: 45.4642
        longitude:
          type: number
          format: float 
          example: 9.1900
        status:
          type: string
          enum:
            - PENDING_APPROVAL
            - ASSIGNED
            - IN_PROGRESS
            - SUSPENDED
            - REJECTED
            - RESOLVED
          example: "ASSIGNED"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "Validation Error"
        message:
          type: string
          description: Detailed error message
          example: "Invalid input data"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authToken
      description: Authentication cookie containing session ID
...